requirements.txt
anyio==4.4.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dnspython==2.7.0
Flask==2.3.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.5
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.43.0
